10-31

+++ jk_parent/jk_entity/src/main/java/com/yaorange/jk/entity/vo/BarChartVO.java
@@ -11,7 +11,7 @@ import java.util.List;
  */
 public class BarChartVO {
     private List<String> titles=new ArrayList<>();
-    private List<Integer> values=new ArrayList<>();
+    private List<Long> values=new ArrayList<>();
 
     public List<String> getTitles() {
         return titles;
@@ -21,11 +21,11 @@ public class BarChartVO {
         this.titles = titles;
     }
 
-    public List<Integer> getValues() {
+    public List<Long> getValues() {
         return values;
     }
 
-    public void setValues(List<Integer> values) {
+    public void setValues(List<Long> values) {
         this.values = values;
     }
 }
